+++ projects/_03_independent_cqc/_07_estimate_filled_posts_by_job_role/fargate/_02_clean.py
@@ -3,10 +3,11 @@ from typing import Final
 import polars as pl
 
 from polars_utils import utils
-from projects._03_independent_cqc._07_estimate_filled_posts_by_job_role.fargate.utils.utils import (
+from projects._03_independent_cqc._07_estimate_filled_posts_by_job_role.fargate.utils.clean_utils import (
     nullify_job_role_count_when_source_not_ascwds,
 )
 from utils.column_names.ind_cqc_pipeline_columns import IndCqcColumns as IndCQC
+from utils.column_values.categorical_column_values import EstimateFilledPostsSource
 
 EXPANDED_ID: Final[str] = "expanded_id"