+++ CHANGELOG.md
@@ -9,6 +9,17 @@ All notable changes to this project will be documented in this file.
 
 
 ### Changed
+
+
+### Fixed
+
+
+## [v2026.03.1] - 21/04/2026
+
+### Added
+- Added function to calculate difference between estimate filled posts and sum of job role estimates.
+
+### Changed
 - Set up imputation step for job role breakdown,  add unit tests and remove functions that are no longer needed.
 
 - Refactored merge job within Estimates by Job Role Pipeline.
@@ -17,7 +28,9 @@ All notable changes to this project will be documented in this file.
 
 - Set up estimates step for job role breakdown.
 
-### Fixed
+- Refactored ASC-WDS filled post calculation code into one file.
+
+- Moved job role utils for cleaning and imputation into their own scripts.
 
 
 ## [v2026.03.0] - 09/04/2026